tests/submission/W3C/domhtml/HTMLObjectElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-object-archive" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object archive="http://www.w3c-test.org/ http://www1.w3c-test.org/" data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<script>
var domtest = async_test("The archive IDL attribute of the object element must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var varchive;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      varchive = testNode.archive;

      assert_equals(varchive,"http://www.w3c-test.org/ http://www1.w3c-test.org/");
       
    domtest.done();
}));

</script>
<p>The archive IDL attribute of the object element must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
